requirements.txt
Flask==3.0.0
Jinja2
markdown
python-frontmatter